refactor: modularise List component
#5
