+++ src/components/grid-events/grid-events.types.ts
@@ -1,0 +1,9 @@
+import { DayEvents, Event } from "../types";
+
+export interface GridEventsProps {
+  events: DayEvents;
+}
+
+export interface EventBlockProps {
+  event: Event;
+}
