+++ src/components/list-icon/list-icon.types.ts
@@ -1,0 +1,3 @@
+export interface ListIconProps {
+  dateTime: Date;
+}
