+++ src/components/list/list.types.tsx
@@ -1,0 +1,5 @@
+import { MonthEvents } from "../types";
+
+export interface ListProps {
+  data: MonthEvents[];
+}
