+++ src/components/month-block/month-block.types.ts
@@ -1,0 +1,7 @@
+export interface MonthBlockProps {
+  dateTime?: Date;
+}
+
+export const MonthBlockDefaultProps = {
+  dateTime: new Date()
+};
